outlook-mcp-server/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "outlook-mcp-server"
version = "0.1.0"
description = "MCP server for Microsoft Outlook / Graph mail (AR Email Management)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.12",
    "msal>=1.28",
    "msgraph-sdk>=1.0",
    "pyjwt[crypto]>=2.8",
    "httpx>=0.27",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[project.scripts]
outlook-mcp-server = "outlook_mcp.server:main"
outlook-mcp-oauth-device = "outlook_mcp.oauth_device:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
